Fix retrieving unicode fields from the server
I happened to have an accented character in a varchar field here that
made mysqldbexport choke with:
UnicodeEncodeError: 'ascii' codec can't encode character u'\xe8' in position 10: ordinal not in range(128)

So we check if the value is a unicode string, and if so we encode it
back to a UTF-8 string.

Works for me(TM) but I didn't run any other test.